feat(D4): settings 9→5 consolidation
- Delete standalone SettingsView (TabView window) + SettingsTab enum
- Merge Audio + Dictation + Translation → AudioDictationSettingsView
- Add About section (version, check-for-updates, link) to GeneralSettingsView
- Remove dead settingsTabToOpen: SettingsTab? from AppState
- Remove SettingsTab.mainSection extension from MainSection
- MeetingSettingsView kept; will surface in D6 Meetings view
- RadioButton + AudioPlayerDelegate helpers moved to AudioDictationSettingsView
